@@ -1,10 +1,10 @@
 name: sign_in_with_apple_example
 description: Demonstrates how to use the sign_in_with_apple plugin.
-publish_to: 'none'
+publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: ">=2.11.0-213.0.dev <2.12.0"
   # For macOS make sure you're on Flutter master or dev channel and
   # that `flutter config` shows `enable-macos-desktop: true`
   flutter: ">=1.12.13+hotfix.8 <2.0.0"
