@@ -76,27 +76,25 @@ class SignInWithApple {
   ///
   /// Throws an [SignInWithAppleNotSupportedException] in case Sign in with Apple is not available (e.g. iOS < 13, macOS < 10.15)
   static Future<AuthorizationCredentialAppleID> getAppleIDCredential({
-    @required List<AppleIDAuthorizationScopes> scopes,
+    required List<AppleIDAuthorizationScopes> scopes,
 
     /// Optional parameters for web-based authentication flows on non-Apple platforms
     ///
     /// This parameter is required on Android.
-    WebAuthenticationOptions webAuthenticationOptions,
+    WebAuthenticationOptions? webAuthenticationOptions,
 
     /// Optional string which, if set, will be be embedded in the resulting `identityToken` field on the [AuthorizationCredentialAppleID].
     ///
     /// This can be used to mitigate replay attacks by using a unique argument per sign-in attempt.
     ///
     /// Can be `null`, in which case no nonce will be passed to the request.
-    String nonce,
+    String? nonce,
 
     /// Data that’s returned to you unmodified in the corresponding [AuthorizationCredentialAppleID.state] after a successful authentication.
     ///
     /// Can be `null`, in which case no state will be passed to the request.
-    String state,
+    String? state,
   }) async {
-    assert(scopes != null);
-
     if (Platform.isAndroid) {
       if (webAuthenticationOptions == null) {
         throw Exception(
@@ -151,8 +149,6 @@ class SignInWithApple {
   static Future<CredentialState> getCredentialState(
     String userIdentifier,
   ) async {
-    assert(userIdentifier != null);
-
     if (!Platform.isIOS &&
         !Platform.isMacOS &&
         Platform.environment['FLUTTER_TEST'] != 'true') {
@@ -183,15 +179,15 @@ class SignInWithApple {
   /// - Android
   ///
   /// In case Sign in with Apple is not available, the returned Future completes with `false`.
-  static Future<bool> isAvailable() {
-    return channel.invokeMethod<bool>('isAvailable');
+  static Future<bool> isAvailable() async {
+    return (await channel.invokeMethod<bool>('isAvailable')) ?? false;
   }
 
   static Future<AuthorizationCredentialAppleID> _signInWithAppleAndroid({
-    @required List<AppleIDAuthorizationScopes> scopes,
-    @required WebAuthenticationOptions webAuthenticationOptions,
-    @required String nonce,
-    @required String state,
+    @required required List<AppleIDAuthorizationScopes> scopes,
+    @required required WebAuthenticationOptions webAuthenticationOptions,
+    @required String? nonce,
+    @required String? state,
   }) async {
     assert(Platform.isAndroid);
 
@@ -203,18 +199,14 @@ class SignInWithApple {
       queryParameters: <String, String>{
         'client_id': webAuthenticationOptions.clientId,
         'redirect_uri': webAuthenticationOptions.redirectUri.toString(),
-        'scope': scopes
-            .map((scope) {
-              switch (scope) {
-                case AppleIDAuthorizationScopes.email:
-                  return 'email';
-                case AppleIDAuthorizationScopes.fullName:
-                  return 'name';
-              }
-              return null;
-            })
-            .where((scope) => scope != null)
-            .join(' '),
+        'scope': scopes.map((scope) {
+          switch (scope) {
+            case AppleIDAuthorizationScopes.email:
+              return 'email';
+            case AppleIDAuthorizationScopes.fullName:
+              return 'name';
+          }
+        }).join(' '),
         // Request `code`, which is also what `ASAuthorizationAppleIDCredential.authorizationCode` contains.
         // So the same handling can be used for Apple and 3rd party platforms
         'response_type': 'code id_token',
@@ -233,6 +225,13 @@ class SignInWithApple {
           'url': uri,
         },
       );
+
+      if (result == null) {
+        throw SignInWithAppleAuthorizationException(
+          code: AuthorizationErrorCode.invalidResponse,
+          message: 'Did receive `null` URL from performAuthorizationRequest',
+        );
+      }
 
       return parseAuthorizationCredentialAppleIDFromDeeplink(Uri.parse(result));
     } on PlatformException catch (exception) {
