@@ -8,12 +8,10 @@ Widget _fallbackBuilder(BuildContext context) => SizedBox.shrink();
 /// A widget for conditionally rendering UI based on whether or not Sign in with Apple is available
 class SignInWithAppleBuilder extends StatefulWidget {
   SignInWithAppleBuilder({
-    Key key,
-    @required this.builder,
+    Key? key,
+    @required required this.builder,
     this.fallbackBuilder = _fallbackBuilder,
-  })  : assert(builder != null),
-        assert(fallbackBuilder != null),
-        super(key: key);
+  }) : super(key: key);
 
   /// A [WidgetBuilder] which will be executed in case Sign in with Apple is available
   final WidgetBuilder builder;
@@ -29,7 +27,7 @@ class SignInWithAppleBuilder extends StatefulWidget {
 
 class _SignInWithAppleBuilderState extends State<SignInWithAppleBuilder> {
   /// Future which will resolve to tell whether or not Sign in with Apple is available
-  Future<bool> _isAvailableFuture;
+  Future<bool>? _isAvailableFuture;
 
   @override
   void initState() {
@@ -43,7 +41,7 @@ class _SignInWithAppleBuilderState extends State<SignInWithAppleBuilder> {
     return FutureBuilder<bool>(
       future: _isAvailableFuture,
       builder: (context, snapshot) {
-        if (snapshot.hasData && snapshot.data) {
+        if (snapshot.hasData && snapshot.data == true) {
           return widget.builder(context);
         }
 
